refactor: extract setup functionality
